==> 1_Text_Classification/03_Data_preprocessing.ipynb <==
import re
import pandas as pd
import numpy as np
import json
from bs4 import BeautifulSoup
from nltk.corpus import stopwords #불용어 제거 모듈
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.preprocessing.text import Tokenizer

DATA_IN_PATH = './data/'
train_data = pd.read_csv(DATA_IN_PATH + 'labeledTrainData.tsv', header=0, delimiter='\t', quoting=3)
print(train_data['review'][0])

def preprocessing(review, remove_stopwords = False):
    # 불용어 제거는 옵션으로 선택 가능
    
    # 1. HTML 태그 제거
    review_text = BeautifulSoup(review, 'html5lib').get_text()
    
    # 2. 영어가 아닌 특수문자를 공백으로 변경
    review_text = re.sub('[^a-zA-Z]', ' ', review_text)
    
    # 3. 대문자를 소문자로 바꾸고 공백 단위로 텍스트를 나눔
    words = review_text.lower().split()
    
    if remove_stopwords:
        # 4. 불용어 제거
        stops = set(stopwords.words('english'))
        # 영어 불용어 불러오기
        words = [w for w in words if not w in stops]
        # 불용어가 아닌 단어로 이루어진 새로운 리스트 생성
        
        # 5. 단어 리스트를 공백을 넣어서 하나의 글로 합침
        clean_review = ' '.join(words)
        
    else: # 불용어를 제거하지 않을 때
        clean_review = ' '.join(words)
    
    return clean_review

clean_train_reviews = []
for review in train_data['review']:
    clean_train_reviews.append(preprocessing(review, remove_stopwords = True))

clean_train_reviews[0]

## 전처리한 데이터에서 각 단어를 인덱스로 벡터화

clean_train_df = pd.DataFrame({'review':clean_train_reviews, 'sentiment':train_data['sentiment']})
clean_train_df.head()
# 전처리한 데이터를 데이터프레임으로 변환

tokenizer = Tokenizer()
tokenizer.fit_on_texts(clean_train_reviews)
text_sequences = tokenizer.texts_to_sequences(clean_train_reviews)
print(text_sequences[0])
# 모델에 따라 입력값이 텍스트가 아닌 각 단어의 인덱스로 되어 있어야 한다.

word_vocab = tokenizer.word_index
#print(word_vocab)
# 각 인덱스가 어떤 단어를 의미하는지 확인이 가능해야 하기 때문에 단어 사전이 필요

print("전체 단어 개수: ", len(word_vocab)+1)

data_configs = {}

data_configs['vocab'] = word_vocab # 단어 사전
data_configs['vocab_size'] = len(word_vocab) # 전체 단어의 개수
# 모델에서 사용하기 위해 따로 저장

## 데이터 길이를 동일하게 해주는 패딩 과정

MAX_SEQUENCE_LENGTH = 174 # 문장 최대 길이, 174인 이유는 EDA 과정에서 나온 중간값

train_inputs = pad_sequences(text_sequences, maxlen=MAX_SEQUENCE_LENGTH, padding='post')
# pad_sequences(데이터, 문장 최대 길이, 패딩 = 'post' : 패딩할 0값을 뒤에 붙임 )
print('Shape of train data :', train_inputs.shape)

## 정답 데이터를 넘파이 배열로 변환

train_labels = np.array(train_data['sentiment'])
print('Shape of label tensor:', train_labels.shape)

# 넘파이 배열로 변환하는 이유는 전처리한 데이터를 저장할 때 넘파이 형태로 저장하기 때문

## 데이터의 저장

TRAIN_INPUT_DATA = 'train_input.npy'
TRAIN_LABEL_DATA = 'train_label.npy'
TRAIN_CLEAN_DATA = 'train_clean.csv'
DATA_CONFIGS = 'data_configs.json'

import os
# 저장하는 디렉토리가 존재하지 않으면 생성
if not os.path.exists(DATA_IN_PATH):
    os.makedirs(DATA_IN_PATH)

# 전처리 된 데이터를 넘파이 형태로 저장
np.save(open(DATA_IN_PATH + TRAIN_INPUT_DATA, 'wb'), train_inputs)
np.save(open(DATA_IN_PATH + TRAIN_LABEL_DATA, 'wb'), train_labels)

# 정제된 텍스트를 csv 형태로 저장
clean_train_df.to_csv(DATA_IN_PATH + TRAIN_CLEAN_DATA, index = False)


# 데이터 사전을 json 형태로 저장
json.dump(data_configs, open(DATA_IN_PATH + DATA_CONFIGS, 'w'), ensure_ascii=False)

## 평가 데이터의 전처리 및 저장

test_data = pd.read_csv(DATA_IN_PATH + "testData.tsv", header=0, delimiter="\t", quoting=3)

clean_test_reviews = []
for review in test_data['review']:
    clean_test_reviews.append(preprocessing(review, remove_stopwords = True))


clean_test_df = pd.DataFrame({'review': clean_test_reviews, 'id': test_data['id']})
test_id = np.array(test_data['id'])

text_sequences = tokenizer.texts_to_sequences(clean_test_reviews)
# 평가 데이터를 전처리 할 때, 학습 데이터에 적용한 토크나이저 객체를 사용해야 한다.
# fit_on_texts의 과정이 없음
# 새롭게 만들면 학습 데이터와 평가 데이터에 대한 각 단어들의 인덱스가 달라져서 모델에 정상적으로 적용 불가.
test_inputs = pad_sequences(text_sequences, maxlen=MAX_SEQUENCE_LENGTH, padding='post')

TEST_INPUT_DATA = 'test_input.npy'
TEST_CLEAN_DATA = 'test_clean.csv'
TEST_ID_DATA = 'test_id.npy'

np.save(open(DATA_IN_PATH + TEST_INPUT_DATA, 'wb'), test_inputs)
np.save(open(DATA_IN_PATH + TEST_ID_DATA, 'wb'), test_id)
clean_test_df.to_csv(DATA_IN_PATH + TEST_CLEAN_DATA, index = False)
