==> 1_Text_Classification/06_RandomForest_with_CountVectorizer.ipynb <==
import pandas as pd
import numpy as np
import os
from sklearn.feature_extraction.text import CountVectorizer

DATA_IN_PATH = './data/'
DATA_OUT_PATH = './output/'
TRAIN_CLEAN_DATA = 'train_clean.csv'
TEST_SIZE = 0.2
RANDOM_SEED = 42

train_data = pd.read_csv(DATA_IN_PATH + TRAIN_CLEAN_DATA)

reviews = list(train_data['review'])
y = np.array(train_data['sentiment'])

## CountVectorizer

vectorizer = CountVectorizer(analyzer = 'word', max_features = 5000)
# 모델에 적용하기 위한 특징 추출 방법의 CountVectorizer
# anlayzer = 분석 단위, max_features = 벡터의 최대 길이
train_data_features = vectorizer.fit_transform(reviews)

train_data_features
# 25,000개의 데이터가 각 5,000개의 특징값을 가지는 벡터로 표현

## 학습, 검증 데이터

from sklearn.model_selection import train_test_split
train_input, eval_input, train_label, eval_label = train_test_split(train_data_features, y, test_size=TEST_SIZE, random_state=RANDOM_SEED)

## 모델

from sklearn.ensemble import RandomForestClassifier

forest = RandomForestClassifier(n_estimators = 100) 
# 랜덤 포레스트 분류기에  100개 의사 결정 트리를 사용한다. (총 100개의 트리를 만들어서 그 결과를 앙상블)

forest.fit( train_input, train_label )

print("Accuracy: %f" % forest.score(eval_input, eval_label))

## 테스트 데이터

TEST_CLEAN_DATA = 'test_clean.csv'

test_data = pd.read_csv(DATA_IN_PATH + TEST_CLEAN_DATA)

test_reviews = list(test_data['review'])
ids = list(test_data['id'])

test_data_features = vectorizer.transform(test_reviews)

if not os.path.exists(DATA_OUT_PATH):
    os.makedirs(DATA_OUT_PATH)
    
result = forest.predict(test_data_features)
# 위에서 만든 랜덤 포레스트 분류기를 통해 예측값을 가져온다.

output = pd.DataFrame( data={"id": ids, "sentiment": result} )
# 판다스 데이터 프레임을 통해 데이터를 구성해서 output에 넣는다.

output.to_csv( DATA_OUT_PATH + "Count_Forest_model.csv", index=False, quoting=3 )

### 캐글에서 0.84264점으로 492등 정도의 결과
